Split big frontend test file into multiple files
This allows playwright to run more tests in parallel
